SemiUpdate

+++ ContactManager.java
@@ -1,5 +1,5 @@
+import java.util.ArrayList;
 import java.util.InputMismatchException;
-import java.util.LinkedList;
 import java.util.Scanner;
 class Contact
 {
@@ -20,76 +20,140 @@ class Contact
 	{
 		this("unknown", 0, phoneNum);
 	}
-	public String getName()
-	{
-		return name;
+	public String getName() { return name; }
+	public int getAge() { return age; }
+	public String getPhoneNum() { return phoneNum; }
+	public void edit(String name, int age, String phoneNum) {
+		this.name = name;
+		this.age = age;
+		this.phoneNum = phoneNum;
 	}
-	public int getAge()
-	{
-		return age;
+	public void edit(String name, String phoneNum) {
+		this.name = name;
+		this.phoneNum = phoneNum;
 	}
-	public String getPhoneNum()
+	public void edit(String phoneNum) {
+		this.phoneNum = phoneNum;
+	}
+	public String toString()
 	{
-		return phoneNum;
+		return name+" "+age+" "+phoneNum;
 	}
 }
 public class ContactManager {
-	private LinkedList<Contact> phoneBook;
+	private ArrayList<Contact> phoneBook;
 	private Scanner scanner;
 	public ContactManager() 
 	{
-		phoneBook = new LinkedList<Contact>();
+		phoneBook = new ArrayList<Contact>();
 		scanner = new Scanner(System.in);
 	}
 	
 	public void showMenu() {
-		
+
+			System.out.println("############ 연락처 관리 ############");
+			System.out.println("1. 연락처 출력");
+			System.out.println("2. 연락처 등록");
+			System.out.println("3. 연락처 삭제");
+			System.out.println("4. 연락처 수정");
+			System.out.println("5. 연락처 검색");
+			System.out.println("6. 출력 설정");
+			System.out.println("7. 끝내기");
+			System.out.print("선택 : ");
+
 	}
 	
 	public void showList() {
-		
+		int i = 1;
+		for(Contact data:phoneBook) {
+			System.out.println(i++ +"."+data);
+		}
 	}
 	
 	public void regist() {
 		
 		String name, phoneNum = null;
 		int age;
-		boolean isDuplicate = true;
+		boolean isDuplicate = false;
+		
 		System.out.print("이름: ");
 		name = scanner.nextLine();
 		System.out.print("나이: ");
 		age = scanner.nextInt();
-		// clearBuffer
-		scanner.nextLine();
-		while(isDuplicate)
-		{
+		scanner.nextLine(); // clear \n
+		
+		do {
 			System.out.print("전화번호: ");
-			phoneNum = scanner.nextLine();
-			isDuplicate = false;
-			for(Contact data : phoneBook)
-			{
-				if(data.getPhoneNum().equals(phoneNum))
-				{
-					isDuplicate = true;
-					System.out.println("전화번호가 다른 연락처와 중복되었습니다. ");
-					break;
-				}
+			try {
+				phoneNum = scanner.nextLine();
+				for(Contact data : phoneBook)
+					if(data.getPhoneNum().equals(phoneNum))
+						throw new Exception("전화번호가 다른 연락처와 중복되었습니다. ");
+				isDuplicate = false;
+			} catch(Exception e) {
+				System.out.println(e.getMessage());
+				isDuplicate = true;
 			}
-		}
+		} while(isDuplicate);
 		phoneBook.add(new Contact(name, age, phoneNum));
-		//System.out.println(phoneBook.get(0).getName() + phoneBook.get(0).getAge() + phoneBook.get(0).getPhoneNum());
 	}
 
-	public void delete() {
-		
+	public void delete() 
+	{
+			int index = 0;
+			System.out.print("삭제할 행번호: ");
+			try {
+				index = scanner.nextInt()-1;
+				scanner.nextLine(); // clear \n
+				if( index > phoneBook.size() || index < 0 )
+					throw new Exception("삭제할 행 번호가 없습니다.");
+			} catch(InputMismatchException ie) {
+				System.out.println("정수만 입력해주세요.");
+			} catch(Exception e) {
+				e.getMessage();
+			}
+			System.out.println( "["+ phoneBook.get(index)+"]"+"가 삭제 되었습니다.");
+			phoneBook.remove(index);
 	}
 	
 	public void edit() {
+
+		boolean rightInput = false;
+		int index = 0;
+		String newNum = null;
 		
+		while(!rightInput)
+		{
+			System.out.print("수정할 행번호 : ");
+			index = scanner.nextInt() - 1;
+			scanner.nextLine(); // clear \n
+			
+			try {
+				System.out.println("수정할 번호를 입력하시오");
+				newNum = scanner.nextLine();
+				
+				rightInput = true;
+			}
+			catch(InputMismatchException e) {
+				System.out.println("번호를 다시 입력해주세요");
+			}
+		}
+		phoneBook.get(index).edit(newNum);
+		System.out.println("수정이 완료되었습니다.");
 	}
 
 	public void search() {
+		String keyword;
+		int i = 1;
+		System.out.print("검색어 : ");
+		keyword =  scanner.nextLine();
 		
+		for(Contact data : phoneBook)
+		{ 
+			if(data.getPhoneNum().contains(keyword))
+				System.out.println(i+". "+data);
+			i++;
+		}
 	}
 	
 	public static void main(String[] args)
@@ -106,6 +170,7 @@ public class ContactManager {
 			{
 				try {
 					act = CM.scanner.nextInt();
+					CM.scanner.nextLine(); // clear \n
 					if(act<1 || act>6) throw new InputMismatchException();
 					rightInput = true;
 				} catch(InputMismatchException e) {
@@ -114,15 +179,22 @@ public class ContactManager {
 					e.printStackTrace();
 				}
 			}
+			System.out.println();
 			switch(act) {
 				case 1:
-					CM.showList();
+					if(CM.phoneBook.size() != 0)
+						CM.showList();
+					else
+						System.out.println("출력할게 없다는 메시지");
 					break;
 				case 2:
 					CM.regist();
 					break;
 				case 3:
-					CM.delete();
+					if(CM.phoneBook.size() != 0)
+						CM.delete();
+					else
+						System.out.println("삭제할게 없다는 메시지");
 					break;
 				case 4:
 					CM.edit();
@@ -136,6 +208,7 @@ public class ContactManager {
 			}
 			rightInput = false;
 			act = 0;
+			System.out.print("\n\n");
 		}
 		CM.scanner.close();
 	}
